Fix async issue with prompts
We added async support to the REST requests in mp_test, however the ui
framework is also async and starts its own event loop. If any
interactive prompts are displayed, there was a conflict in that a second
event loop tries to start while the main event loop is still running.
The solution we have is to add the 'nest_asyncio' module, which allows
event loops to be nested.