Do not save project.path to project config file
While we want to know a project's path, we should store it when we read
the config.json file, instead of reading it from the config.json file.
This is because if a project moves, the path in config.json becomes out
of date, but we already have the updated path, because that is used to
read the config.json file in the first place.